size_t fix

+++ asafov.andrew/P3/matrixAutoInitialization.cpp
@@ -7,7 +7,7 @@ void asafov::matrixAutoInitialization(int* mtx, char* input, size_t& n, size_t& 
   std::ifstream fin(input);
   fin >> n;
   fin >> m;
-  for (int i = 0; i < n * m; i++) {
+  for (size_t i = 0; i < n * m; i++) {
     fin >> mtx[i];
     if (fin.fail() || fin.eof()) {
       throw std::logic_error("err");
